Feat : API 호출 관계 정리, 모델 재구성, movies.views내에 기능 추가
