+++ pjt-django/movies/templates/movies/index.html
@@ -14,7 +14,7 @@
         <div class="img-container card bg-dark text-white" style="width: 18rem;">          
           <img src={{ "https://www.themoviedb.org/t/p/w300_and_h450_bestv2"|add:movie.poster_path }} class="card-img img-thumbnail" alt="...">
           <div class="card-img-overlay d-flex align-items-end">
-            <h5 class="text card-title">{{ movie.title }}</h5>     
+            <h5 class="text card-title">{{ movie.title }}</h5>
             <button class="play-btn" data-bs-toggle="modal" data-bs-target="#exampleModal{{movie.pk}}">Play</button>     
           </div>
         </div>
